+++ utils.js
@@ -1,0 +1,7 @@
+let date = new Date();
+export let data = (date.getHours() + ':' + date.getMinutes())
+
+
+export const getRandom = (num) => {
+    return Math.ceil(Math.random() * num)
+}
