+++ variables.js
@@ -1,3 +1,5 @@
+import {getRandom} from './utils.js'
+
 export const HIT = {
     head: 30,
     body: 25,
@@ -8,9 +10,6 @@ export const HIT = {
 export const ATTACK = ['head', 'body', 'foot'];
 
 
-let date = new Date();
-export let data = (date.getHours() + ':' + date.getMinutes())
-/*
 export const player1 = {
     player: 1,
     name: 'SCORPION',
@@ -39,4 +38,20 @@ export const player2 = {
     renderHP,
     elHP,
 }
-*/
+
+export function elHP () {
+    return document.querySelector(`.player${this.player} .life`);
+}
+
+export function renderHP() {
+    this.elHP().style.width = this.hp + '%'
+}
+
+export function changeHP () {
+    this.hp -= getRandom(20)
+    if (this.hp <= 0) {
+        this.hp = 0
+    }
+    this.renderHP()
+    return this.hp
+}
